feat: User Entity 설계
