+++ src/main/java/project/valuetogether/domain/entity/Gender.java
@@ -1,0 +1,5 @@
+package project.valuetogether.domain.entity;
+
+public enum Gender {
+    남성,여성
+}
