+++ src/main/java/project/valuetogether/domain/entity/Role.java
@@ -1,0 +1,5 @@
+package project.valuetogether.domain.entity;
+
+public enum Role {
+    ROLE_USER, ROLE_PROPOSER, ROLE_ADMIN
+}
