+++ src/main/java/project/valuetogether/domain/entity/User.java
@@ -1,0 +1,86 @@
+package project.valuetogether.domain.entity;
+
+import com.sun.istack.NotNull;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+@Data
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "tbl_user")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull
+    private String email;
+
+    @NotNull
+    private String password;
+
+    @NotNull
+    private String name;
+
+    @NotNull
+    @Column(length = 13)
+    private String phoneNumber;
+
+    @NotNull
+    private String address;
+
+    @NotNull
+    @Enumerated(EnumType.STRING)
+    private Gender gender;
+
+    @NotNull
+    private String nickname;
+
+    @NotNull
+    private String birthYear;
+
+    @NotNull
+    @Enumerated(EnumType.STRING)
+    private Role role = Role.ROLE_USER;
+
+
+
+    @Builder
+    public User(String email, String password, String name, String phoneNumber, String address, Gender gender, String nickname, String birthYear) {
+        this.email = email;
+        this.password = password;
+        this.name = name;
+        this.phoneNumber = phoneNumber;
+        this.address = address;
+        this.gender = gender;
+        this.nickname = nickname;
+        this.birthYear = birthYear;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public Role getRole() {
+        return role;
+    }
+}
