+++ src/main/resources/application.yml
@@ -6,9 +6,9 @@ spring:
     driver-class-name: org.h2.Driver
 
   jpa:
-    hibernate.hbm2ddl.auto: update
-    #    hibernate:
-    #      ddl-auto: create
+#    hibernate.hbm2ddl.auto: update
+    hibernate:
+      ddl-auto: create
     properties:
       hibernate:
         # show_sql: true
